+++ nis/nss_compat/compat-pwd.c
@@ -677,7 +677,7 @@ getpwent_next_file_plususer (struct passwd *result, char *buffer,
 	  != YPERR_SUCCESS)
 	return NSS_STATUS_TRYAGAIN;
       p = strncpy (buffer, outval,
-		   buflen < outvallen ? buflen : outvallen);
+		   buflen < (size_t) outvallen ? buflen : (size_t) outvallen);
       free (outval);
       while (isspace (*p))
 	p++;
